Fix some @jerome1337 comments
