--- models/Cell/Alignment.go
@@ -1,8 +1,0 @@
-package models
-
-//Alignment : Represent a json cell style model
-type Alignment struct {
-	Horizontal string `json:"horizontal,omitempty"`
-	Vertical   string `json:"vertical,omitempty"`
-	WrapText   bool   `json:"wrap_text,omitempty"`
-}
