--- models/Cell/Protection.go
@@ -1,7 +1,0 @@
-package models
-
-//Protection : Represent a json cell style model
-type Protection struct {
-	Hidden bool `json:"hidden,omitempty"`
-	Locked bool `json:"locked,omitempty"`
-}
