+++ models/Header.go
@@ -1,6 +1,6 @@
 package models
 
-// Header : Represent a json model Header
+// Header represent a json model Header
 type Header struct {
 	Orientation string    `json:"orientation,omitempty"` // Can be "col" or "row", default "col"
 	Name        string    `json:"name"`
