+++ models/Item.go
@@ -3,6 +3,6 @@ package models
 // Item represent json model Item
 type Item struct {
 	Mode                    string  `json:"mode"`
-	StartingCellCoordonates string  `json:"startingCellCoordonates,omitempty"`
+	StartingCellCoordonates string  `json:"starting_cell_coordonates,omitempty"`
 	Datas                   []*Data `json:"datas"`
 }
